requirements.txt
djangorestframework 
djoser
djangorestframework-simplejwt
django-cors-headers